Get rid of most of the code and try to work with what Flask already provides.
This removes the middleware, basically everything and only adds
syntactic sugar for request.environ['wsgi.websocket'], made available
like other request-bound locals like g or session.

This causes url routing to work a lot better without having to
reimplement half of it in the extension itself.

Things that still need to be addressed are the superflous empty-string
return to avoid exceptions caused by a view not sending a response. This
may get cleverly patched away later on.

Another thing that'd be nice to get working is a good plan for url_for,
possibly adding the ``ws://``-prefix if a URL is declared
websocket-only. Using a decorator there could fix the return-value issue
as well.